keep the 'records' dialog shown in the viewport
- after scrolling the 'deduplication' table & clicking a cell in the
'Count' columnn, display the 'records' dialog in the viewport
