+++ src/css/plugin-layout.module.css
@@ -120,8 +120,8 @@
   position: absolute;
   z-index: 10;
   top: 10%;
-  left: 25%;
-  width: 50%;
+  left: 17.5%;
+  width: 35%;
   height: 80%;
   max-height: 500px;
   flex-shrink: 0;
